requirements.txt
sqlalchemy
accelerate==1.12.0
adbutils==2.11.0
aistudio-sdk==0.3.8
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bce-python-sdk==0.9.55
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
fastapi==0.124.0
filelock==3.20.0
facebook-wda==1.4.6
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.12.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.2
macholib==1.16.4
MarkupSafe==3.0.3
matplotlib==3.10.7
modelscope==1.32.0
mpmath==1.3.0
networkx==3.6
numpy==2.2.6
onnxruntime==1.23.2
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.10
pandas==2.3.3
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
prettytable==3.17.0
protobuf==6.33.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pynput==1.8.1
pyobjc-core==12.1; sys_platform == 'darwin'
pyobjc-framework-ApplicationServices==12.1; sys_platform == 'darwin'
pyobjc-framework-Cocoa==12.1; sys_platform == 'darwin'
pyobjc-framework-CoreText==12.1; sys_platform == 'darwin'
pyobjc-framework-Quartz==12.1; sys_platform == 'darwin'
pyparsing==3.2.5
pypdfium2==5.1.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
rapidocr-onnxruntime==1.2.3
regex==2025.11.3
requests==2.32.5
retry2==0.9.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scipy==1.16.3
selenium==4.38.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-websocket==0.12.2
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uiautomator2==3.5.0
ujson==5.11.0
ultralytics==8.3.235
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==25.10.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
duckdb==1.4.3
python-multipart==0.0.20
pywinauto==0.6.9